% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_entry.glm.R
\name{generate_entry.glm}
\alias{generate_entry.glm}
\title{Create a log entry for GLM objects}
\usage{
\method{generate_entry}{glm}(object)
}
\arguments{
\item{object}{The model object of the model for which an entry is desired}
}
\value{
Named list of values to be used as a line entry in the model log.
}
\description{
Create a log entry for GLM objects
}

