% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_entry.R
\name{generate_entry}
\alias{generate_entry}
\title{Helper function to create log entry object}
\usage{
generate_entry(object)
}
\arguments{
\item{object}{The model object of the model to be logged}
}
\value{
Named list of values to be used as a line entry in the model log.
}
\description{
Helper function to create log entry object
}

