% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_add.glm.R
\name{log_add.glm}
\alias{log_add.glm}
\title{Add an entry to the log}
\usage{
\method{log_add}{glm}(model_obj, log_path)
}
\arguments{
\item{model_obj}{The model object of the model to be logged}

\item{log_path}{File path to the log file. Default is the current working
directory.}
}
\value{
A long entry is recorded in the specified file, and the modelobject
is passed through.
}
\description{
Add an entry to the log
}

