% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_add.R
\name{log_add}
\alias{log_add}
\title{Add an entry to the log}
\usage{
log_add(model_obj, log_path)
}
\arguments{
\item{model_obj}{The model object of the model to be logged}

\item{log_path}{File path to the log file. Default is the current working
directory.}
}
\value{
A log file named "model_log.txt is created in the directory"
}
\description{
Add an entry to the log
}

