% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_entry.R
\name{log_entry}
\alias{log_entry}
\title{Log Entry}
\usage{
log_entry(object, conn)
}
\arguments{
\item{object}{The model object whose info we wish to log}

\item{conn}{The connection to the log}
}
\description{
Writes a log entry to the log
}

