% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_init.R
\name{log_init}
\alias{log_init}
\title{Log file creation}
\usage{
log_init(path = getwd())
}
\arguments{
\item{path}{File path to the log file. Default is the current working
directory.}
}
\value{
A log file named "model_log.csv" is created in the directory
}
\description{
Initialize the log file
}

