% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model.glm.R
\name{parse_model.glm}
\alias{parse_model.glm}
\title{Extract relevant model info for GLM objects}
\usage{
\method{parse_model}{glm}(object)
}
\arguments{
\item{object}{The model object of the model to extract info from}
}
\value{
Named list of values to be used as a line entry in the model log.
}
\description{
Extract relevant model info for GLM objects
}

