% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_model.R
\name{parse_model}
\alias{parse_model}
\title{Helper function to extract relevant info from model objects}
\usage{
parse_model(object)
}
\arguments{
\item{object}{The model object of the model to be logged}
}
\value{
Named list of values to be used as part of a log entry
}
\description{
Helper function to extract relevant info from model objects
}

