% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_add.R
\name{log_add}
\alias{log_add}
\title{log_add is a generic function that records the fitting of a model in the
provided log system.}
\usage{
log_add(model_obj, log_path)
}
\arguments{
\item{model_obj}{The model object of the model to be logged}

\item{log_path}{File path to the log file. Default is the current working
directory.}
}
\value{
A log entry is recorded in the specified file, and the modelobject
is passed through.
}
\description{
log_add is a generic function that records the fitting of a model in the
provided log system.
}

